requirements.txt
requests
pandas
tomark
fuzzywuzzy
python-Levenshtein
rich
pyyaml